requirements.txt
django~=4.1
pillow~=10.3.0
numpy~=1.26.4
tensorflow~=2.16.1